Changed VisitorSessionContainer to use live data
